requirements.txt
autopep8
conan
gitpython